refactor(context): migrate collect_retrieved_context to zeph-agent-context
Move RetrievedContext and collect_retrieved_context from
zeph-core::agent::context::retrieved into zeph-agent-context::retrieved
(feature-gated behind self-check).

zeph-core::quality::pipeline now re-exports RetrievedContext via
pub use so all existing consumer paths remain valid without changes.
The zeph-agent-context crate gains no new dependency on zeph-core.

Activates zeph-agent-context/self-check from zeph-core's self-check
feature flag (was empty before).

Closes #3529
